Allow clients to register themselves for debugging
A client can send a register_process command with some tags at any time (for
backward-compatiblity).

Stagger associates these tags to the client. They could be used to append to
tagged stats but right now they are only used for debug.

This helps debugging which client id corresponds to which process.